generativeaichatbot_project/requirements.txt
Django
djangorestframework
google-cloud-storage
transformers
torch
pymilvus
gunicorn